did a bit of work with extracting headers from http req

+++ c-server/extract_header.h
@@ -1,1 +1,1 @@
-int64_t extract_header(char * reqContent, char * headerName, char ** output);
+size_t extract_header(char * reqContent, char * headerName, char ** output);
